+++ src/Logikfabrik.Umbraco.Jet/Logging/LogEntry.cs
@@ -1,0 +1,67 @@
+﻿// <copyright file="LogEntry.cs" company="Logikfabrik">
+//   Copyright (c) 2016 anton(at)logikfabrik.se. Licensed under the MIT license.
+// </copyright>
+
+namespace Logikfabrik.Umbraco.Jet.Logging
+{
+    using System;
+
+    /// <summary>
+    /// The <see cref="LogEntry" /> class.
+    /// </summary>
+    public class LogEntry
+    {
+        /// <summary>
+        /// Initializes a new instance of the <see cref="LogEntry"/> class.
+        /// </summary>
+        /// <param name="type">The type.</param>
+        /// <param name="message">The message.</param>
+        public LogEntry(LogEntryType type, string message)
+            : this(type, message, null)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="LogEntry"/> class.
+        /// </summary>
+        /// <param name="type">The type.</param>
+        /// <param name="message">The message.</param>
+        /// <param name="exception">The exception.</param>
+        /// <exception cref="ArgumentException">Thrown if <paramref name="message" /> is <c>null</c> or white space.</exception>
+        public LogEntry(LogEntryType type, string message, Exception exception)
+        {
+            if (string.IsNullOrWhiteSpace(message))
+            {
+                throw new ArgumentException("Message cannot be null or white space.", nameof(message));
+            }
+
+            Type = type;
+            Message = message;
+            Exception = exception;
+        }
+
+        /// <summary>
+        /// Gets the type.
+        /// </summary>
+        /// <value>
+        /// The type.
+        /// </value>
+        public LogEntryType Type { get; }
+
+        /// <summary>
+        /// Gets the message.
+        /// </summary>
+        /// <value>
+        /// The message.
+        /// </value>
+        public string Message { get; }
+
+        /// <summary>
+        /// Gets the exception.
+        /// </summary>
+        /// <value>
+        /// The exception.
+        /// </value>
+        public Exception Exception { get; }
+    }
+}