+++ src/Logikfabrik.Umbraco.Jet/MemberTypeSynchronizer.cs
@@ -9,6 +9,7 @@ namespace Logikfabrik.Umbraco.Jet
     using Data;
     using global::Umbraco.Core.Models;
     using global::Umbraco.Core.Services;
+    using Logging;
 
     /// <summary>
     /// The <see cref="MemberTypeSynchronizer" /> class. Synchronizes model types annotated using the <see cref="MemberTypeAttribute" />.
@@ -21,15 +22,17 @@ namespace Logikfabrik.Umbraco.Jet
         /// <summary>
         /// Initializes a new instance of the <see cref="MemberTypeSynchronizer" /> class.
         /// </summary>
+        /// <param name="logService">The log service.</param>
         /// <param name="memberTypeService">The member type service.</param>
         /// <param name="typeResolver">The type resolver.</param>
         /// <param name="typeRepository">The type repository.</param>
         /// <exception cref="ArgumentNullException">Thrown if <paramref name="memberTypeService" />, or <paramref name="typeResolver" /> are <c>null</c>.</exception>
         public MemberTypeSynchronizer(
+            ILogService logService,
             IMemberTypeService memberTypeService,
             ITypeResolver typeResolver,
             ITypeRepository typeRepository)
-            : base(typeRepository)
+            : base(logService, typeRepository)
         {
             if (memberTypeService == null)
             {